Moved compare components to compare folder
