+++ client/src/compare/ComparePlayerHeadshots.js
