+++ client/src/compare/index.js
