+++ spec/client/compare.js
@@ -1,0 +1,10 @@
+import { expect } from 'chai';
+
+describe('compare view', () => {
+
+  it ('should only receive an array of up to two players', (done) => {
+    expect(1+1).to.equal(2);
+    done()
+  });
+
+});
