+++ spec/client/test.jsx
@@ -5,7 +5,6 @@ import { shallow, mount, render } from 'enzyme';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
 
-
 import reducers from '../../client/src/reducers/root_reducer';
 
 import App from '../../client/src/App';
@@ -23,12 +22,4 @@ describe("A suite", function() {
     );
     expect(wrapper.contains(<div className="app-container" />)).to.equal(true);
   });
-
-  // it("contains spec with an expectation", function() {
-  //   expect(shallow(<App />).is('.foo')).to.equal(true);
-  // });
-  //
-  // it("contains spec with an expectation", function() {
-  //   expect(mount(<App />).find('.foo').length).to.equal(1);
-  // });
 });
